--- app/src/main/java/com/sun/audiorecorder/ui/ChangeLanguage.java
@@ -1,30 +1,0 @@
-package com.sun.audiorecorder.ui;
-
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.sun.audiorecorder_01.R;
-
-
-/**
- * A simple {@link Fragment} subclass.
- */
-public class ChangeLanguage extends Fragment {
-
-    public ChangeLanguage() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_change_language, container, false);
-    }
-
-}
