+++ app/src/main/java/com/sun/audiorecorder/ui/HomeFragment.java
@@ -13,7 +13,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
-import android.widget.RadioButton;
 
 import com.sun.audiorecorder_01.R;
 
@@ -21,13 +20,8 @@ import com.sun.audiorecorder_01.R;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class HomeFragment extends Fragment implements View.OnClickListener {
-
+public class HomeFragment extends Fragment {
     private AlertDialog.Builder mbuilder;
- //   private ChangeBackgroud mChangeBackgroud;
-    private ImageButton mImageButtonListAudioRecover, mImageButtonPlay, mImageButtonStop;
-    private OpenListAudioRecorverFragmentCallback mOpenListAudioRecorverFragmentCallback;
-    protected boolean check=true;
     private static HomeFragment sInstance;
 
     public static synchronized HomeFragment getInstance() {
@@ -48,58 +42,7 @@ public class HomeFragment extends Fragment implements View.OnClickListener {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        initView();
-        initClickListeners();
         setHasOptionsMenu(true);
-    }
-
-    @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        if (getActivity() instanceof OpenListAudioRecorverFragmentCallback) {
-            mOpenListAudioRecorverFragmentCallback = (OpenListAudioRecorverFragmentCallback) getActivity();
-        }
-    }
-
-    private void initClickListeners() {
-        mImageButtonListAudioRecover.setOnClickListener(this);
-        mImageButtonPlay.setOnClickListener(this);
-        mImageButtonStop.setOnClickListener(this);
-    }
-
-    private void initView() {
-        mImageButtonListAudioRecover = getView().findViewById(R.id.imagebutton_ListAudioRecover);
-        mImageButtonPlay = getView().findViewById(R.id.imagebutton_play);
-        mImageButtonStop = getView().findViewById(R.id.imagebutton_stop);
-    }
-
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.imagebutton_ListAudioRecover:
-                mOpenListAudioRecorverFragmentCallback.OpenListAudioRecorverFragmentCallback();
-                break;
-            case R.id.imagebutton_play:
-                initEventClickListeners();
-                break;
-            case R.id.imagebutton_stop:
-                mOpenListAudioRecorverFragmentCallback.OpenListAudioRecorverFragmentCallback();
-                break;
-        }
-    }
-
-    private void initEventClickListeners() {
-        if (check) {
-            check=false;
-            mImageButtonPlay.setBackgroundResource(R.drawable.btn_pause);
-        } else {
-            check=true;
-            mImageButtonPlay.setBackgroundResource(R.drawable.btn_play);
-        }
-    }
-
-    private void initDiaLogChangeBackgroundListeners() {
-
     }
 
     @Override
@@ -111,25 +54,24 @@ public class HomeFragment extends Fragment implements View.OnClickListener {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.item_Change_Language:
-                OpenDiaLogChangeLanguage();
+                openDiaLogChangeLanguage();
                 break;
             case R.id.item_Change_Background:
-                OpenDiaLogChangeBackground();
+                openDiaLogChangeBackground();
                 break;
         }
         return super.onOptionsItemSelected(item);
     }
 
-    private void OpenDiaLogChangeBackground() {
+    private void openDiaLogChangeBackground() {
         mbuilder = new AlertDialog.Builder(getActivity());
         View view = getLayoutInflater().inflate(R.layout.fragment_change_backgroud, null);
         mbuilder.setView(view);
         mbuilder.create();
         mbuilder.show();
-        initDiaLogChangeBackgroundListeners();
     }
 
-    private void OpenDiaLogChangeLanguage() {
+    private void openDiaLogChangeLanguage() {
         mbuilder = new AlertDialog.Builder(getActivity());
         View view = getLayoutInflater().inflate(R.layout.fragment_change_language, null);
         mbuilder.setView(view);
@@ -137,10 +79,5 @@ public class HomeFragment extends Fragment implements View.OnClickListener {
         mbuilder.show();
     }
 
-
-
-    public interface OpenListAudioRecorverFragmentCallback {
-        void OpenListAudioRecorverFragmentCallback();
-    }
-
 }
+
