+++ app/src/main/res/layout/fragment_home.xml
@@ -6,70 +6,4 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".ui.HomeFragment">
 
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="150dp"
-        android:layout_height="160dp"
-        android:layout_marginStart="@dimen/marginStart_8"
-        android:layout_marginTop="@dimen/height_30"
-        android:layout_marginEnd="@dimen/margin_end_112"
-        android:src="@drawable/ic_mic_black"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:ignore="MissingConstraints" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="@dimen/width_162"
-        android:layout_height="@dimen/height_50"
-        android:layout_marginStart="@dimen/marginStart_8"
-        android:layout_marginTop="@dimen/margintop_16"
-        android:layout_marginEnd="@dimen/marginend_8"
-        android:text="00:00:00"
-        android:textColor="@color/colorBlack"
-        android:textSize="@dimen/textsize_40"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.557"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageView"
-        tools:ignore="MissingConstraints" />
-
-    <ImageButton
-        android:id="@+id/imagebutton_ListAudioRecover"
-        android:layout_width="@dimen/width_62"
-        android:layout_height="@dimen/height_62"
-        android:layout_marginTop="@dimen/margintop_48"
-        android:layout_marginStart="@dimen/marginstart_40"
-        android:background="@drawable/btn_list_audio_recover"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.055"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView"
-        tools:ignore="MissingConstraints"/>
-
-    <ImageButton
-        android:id="@+id/imagebutton_stop"
-        android:layout_width="@dimen/width_62"
-        android:layout_height="@dimen/height_62"
-        android:layout_marginTop="@dimen/margintop_48"
-        android:layout_marginEnd="@dimen/marginend_40"
-        android:background="@drawable/btn_stop"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.972"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
-
-    <ImageButton
-        android:id="@+id/imagebutton_play"
-        android:layout_width="@dimen/width_62"
-        android:layout_height="@dimen/height_62"
-        android:layout_marginTop="@dimen/margintop_48"
-        android:background="@drawable/btn_play"
-        app:layout_constraintEnd_toStartOf="@+id/imagebutton_stop"
-        app:layout_constraintStart_toEndOf="@+id/imagebutton_ListAudioRecover"
-        app:layout_constraintTop_toBottomOf="@+id/textView"
-        tools:ignore="MissingConstraints" />
-
-
 </android.support.constraint.ConstraintLayout>
